[FE] style: 회원가입 페이지 컴포넌트 분리
